requirements-lock.txt
# ===================================================================
# AI Calendar Assistant - Complete Lock File
# This contains ALL packages including transitive dependencies
# Generated on: 2025-07-16
# Python Version: 3.13.5
# Use this for exact reproduction of the environment
# ===================================================================

aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
asyncio==3.4.3
attrs==25.3.0
av==14.4.0
azure-ai-agents==1.1.0b4
azure-ai-projects==1.0.0b12
azure-core==1.32.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-cosmos==4.9.0
azure-identity==1.23.0
azure-monitor-opentelemetry==1.6.0
azure-monitor-opentelemetry-exporter==1.0.0b39
azure-storage-blob==12.25.1
certifi==2025.7.9
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudevents==1.12.0
colorama==0.4.6
cryptography==45.0.5
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
fixedint==0.1.6
frozenlist==1.7.0
google-crc32c==1.7.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
hyperframe==6.1.0
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.6.1
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
lazy-object-proxy==1.11.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
microsoft-kiota-abstractions==1.9.4
microsoft-kiota-authentication-azure==1.9.4
microsoft-kiota-http==1.9.4
microsoft-kiota-serialization-form==1.9.4
microsoft-kiota-serialization-json==1.9.4
microsoft-kiota-serialization-multipart==1.9.4
microsoft-kiota-serialization-text==1.9.4
more-itertools==10.7.0
msal==1.32.3
msal-extensions==1.3.1
msgraph-core==1.3.5
msgraph-sdk==1.37.0
msrest==0.7.1
multidict==6.6.3
nest-asyncio==1.6.0
numpy==2.3.1
oauthlib==3.3.1
openai==1.95.1
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.31.0
opentelemetry-instrumentation==0.52b0
opentelemetry-instrumentation-asgi==0.52b0
opentelemetry-instrumentation-dbapi==0.52b0
opentelemetry-instrumentation-django==0.52b0
opentelemetry-instrumentation-fastapi==0.52b0
opentelemetry-instrumentation-flask==0.52b0
opentelemetry-instrumentation-httpx==0.52b0
opentelemetry-instrumentation-logging==0.52b0
opentelemetry-instrumentation-psycopg2==0.52b0
opentelemetry-instrumentation-requests==0.52b0
opentelemetry-instrumentation-urllib==0.52b0
opentelemetry-instrumentation-urllib3==0.52b0
opentelemetry-instrumentation-wsgi==0.52b0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
opentelemetry-util-http==0.52b0
orjson==3.10.18
packaging==25.0
parse==1.20.2
pathable==0.4.4
pip==25.1.1
prance==25.4.8.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
psycopg2==2.9.10
pybars4==0.9.13
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pylibsrtp==0.12.0
PyMeta3==0.5.1
pyOpenSSL==25.1.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.2
rpds-py==0.26.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scipy==1.16.0
semantic-kernel==1.34.0
sentry-sdk==2.32.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
std-uritemplate==2.0.5
tqdm==4.67.1
typer==0.16.0
typing==3.7.4.3
typing_extensions==4.14.1
typing-inspection==0.4.1
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.1
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
